move ApiError to errors.js
