+++ lib/api.js
@@ -1,10 +1,11 @@
 var _ = require('lodash');
-var util = require('util');
+
 var url = require('url');
 var request = require('./request');
 var Promise = require('q').Promise;
 var packageJson = require('../package.json');
 var query = require('qs');
+var anxErrors = require('./errors');
 
 function _hasValue(value) {
 	return !(_.isNull(value) || _.isUndefined(value));
@@ -12,6 +13,12 @@ function _hasValue(value) {
 
 function _isInteger(value) {
 	return parseInt(value, 10) === +value;
+}
+
+function _normalizeOpts(opts) {
+	return _.isString(opts) ? {
+		uri: opts
+	} : _.cloneDeep(opts) || {};
 }
 
 function Api(config) {
@@ -23,123 +30,8 @@ function Api(config) {
 	});
 }
 
-function _normalizeOpts(opts) {
-	return _.isString(opts) ? {
-		uri: opts
-	} : _.cloneDeep(opts) || {};
-}
-
-Api.ApiError = function() {
-	Error.apply(this);
-	Error.captureStackTrace(this, this.constructor);
-
-	this.name = 'ApiError';
-	this.defaultMessage = '';
-
-	var id, code, message, description;
-	switch (arguments.length) {
-		case 1: // new ApiError(response)
-			var err = arguments[0];
-			if (_.isObject(err)) {
-
-				// Traverse through general API JSON Response
-				if (err.body && err.body.response) {
-					err = err.body.response; // err is raw api response
-				} else if (err.response) {
-					err = err.response; // err is body
-				}
-
-				// Extract values from object - duck type check
-				if (err.hasOwnProperty('error_id')) { // Api Response
-					id = err.error_id;
-					code = err.error_code;
-					message = err.error;
-					description = err.error_description;
-				} else if (err.hasOwnProperty('id')) { // Simple Object
-					id = err.id;
-					code = err.code;
-					message = err.message;
-					description = err.description;
-				} else {
-					message = err;
-				}
-			} else {
-				// Even if not string - let the argument be message like default Error behavior
-				message = err;
-			}
-			break;
-		case 2: // new ApiError(id, message)
-			id = arguments[0];
-			message = arguments[1];
-			break;
-		case 4:
-			id = arguments[0];
-			code = arguments[1];
-			message = arguments[2];
-			description = arguments[3];
-			break;
-		case 3: // new ApiError(id, code, message)
-			id = arguments[0];
-			code = arguments[1];
-			message = arguments[2];
-			break;
-	}
-
-	this.id = id;
-	this.code = code;
-	this.message = (message || this.defaultMessage);
-	this.description = (description || '');
-};
-util.inherits(Api.ApiError, Error);
-
-Api.NotAuthorizedError = function() {
-	this.defaultMessage = 'Authorization failed';
-	Api.ApiError.apply(this, arguments);
-	this.name = 'NotAuthorizedError';
-};
-util.inherits(Api.NotAuthorizedError, Api.ApiError);
-
-// NotAuthenticated extends NotAuthorized for backwards compatibility
-Api.NotAuthenticatedError = function() {
-	this.defaultMessage = 'Authentication failed';
-	Api.NotAuthorizedError.apply(this, arguments);
-	this.name = 'NotAuthenticatedError';
-};
-util.inherits(Api.NotAuthenticatedError, Api.NotAuthorizedError);
-
-// Build error from root response
-// https://wiki.appnexus.com/display/adnexusdocumentation/API+Semantics#APISemantics-Errors
-Api.buildError = function(res) {
-	var ErrorType = Api.ApiError;
-	if (res) {
-		var statusCode = res.statusCode;
-		var errorId;
-		if (res && res.body && res.body.response && res.body.response) {
-			errorId = res.body.response.error_id;
-		}
-
-		if (statusCode || errorId) {
-			// Differentiating Authentication vs Authorization [http://stackoverflow.com/a/6937030/2483105]
-			if (statusCode === 401 || errorId === 'NOAUTH') {
-				ErrorType = Api.NotAuthenticatedError;
-			} else if (statusCode === 403 || errorId === 'UNAUTH') {
-				ErrorType = Api.NotAuthorizedError;
-			}
-		}
-	}
-	return new ErrorType(res);
-};
-
-Api.TargetError = function() {
-	Api.ApiError.apply(this, arguments);
-	this.name = 'TargetError';
-};
-util.inherits(Api.TargetError, Api.ApiError);
-
-
-Api.statusOk = function _statusOk(body) {
-	return body && body.response && body.response.status === 'OK';
-};
+// Bind error types on the Api namespace
+anxErrors.extend(Api);
 
 Api.prototype._request = function _request(opts) {
 	var _self = this;
